432: Fix bugs: endpointIndex initialization and calling the template arg URIParser

+++ network/http_parser/HTTPParsingInfo.h
@@ -1,8 +1,9 @@
 #pragma once
 
-#include "BasicResult.h"
 #include <string_view>
 #include <array>
+
+#include "BasicResult.h"
 
 namespace net::HTTP {
 
@@ -12,7 +13,7 @@ using Uri = std::string_view;
 using Path = std::string_view;
 using Payload = std::string_view;
 using Params = std::array<std::string_view, (size_t)MAX_PARAM_COUNT>;
-using EndpointIndex = size_t;
+using EndpointIndex = int64_t;
 
 enum class Error {
     UNKNOWN = 0,
